+++ src/Geometry/Factories/LineFactory.php
@@ -31,9 +31,9 @@ class LineFactory implements DrawableFactoryInterface
     /**
      * {@inheritdoc}
      *
-     * @see DrawableFactoryInterface::init()
+     * @see DrawableFactoryInterface::create()
      */
-    public static function init(null|Closure|DrawableInterface $init = null): self
+    public static function create(null|Closure|DrawableInterface $init = null): self
     {
         return new self($init);
     }
@@ -41,9 +41,9 @@ class LineFactory implements DrawableFactoryInterface
     /**
      * {@inheritdoc}
      *
-     * @see DrawableFactoryInterface::create()
+     * @see DrawableFactoryInterface::drawable()
      */
-    public function create(): DrawableInterface
+    public function drawable(): DrawableInterface
     {
         return $this->line;
     }